fix(render): review fixes for CID-keyed CFF (byb-8b9.8)
- Cap FDArray at 256 and Subrs INDEX at 40000, matching sfnt's
  maxNumFontDicts/maxNumSubroutines: the pre-parse gate ran before
  sfnt's own caps, so a tiny file declaring 65535 FDs sharing one
  65535-entry Subrs INDEX allocated numFD x numSubrs slice headers
  (measured in the GB) before any budget fired.
- Mirror sfnt's parsePrivateDICT edge exactly: skip a zero-length
  Private, refuse a non-empty one at offset <= 0.
- Fold showText's duplicated advance arithmetic into one shared
  closure so the Tc/Tz terms exist once and are pinned by
  TestTextSpacingParamsExact on both paths.
- Narrow Font.W to map[uint16]float64: showText builds uint16 CIDs.
- Pin the /DW branch (DW 500 halves an unmapped CID's advance) and
  cffFDSelect's unsorted-tolerant bisection (a first-match linear
  scan diverges at gids 20/39 of the new fixture).

+++ internal/render/oracle_test.go
@@ -470,7 +470,7 @@ func TestCIDKeyedCFFTextAgreesWithPdftoppm(t *testing.T) {
 			return Font{}, false
 		}
 		return Font{Program: oracleCIDCFF(), Type0: true, DW: 1000,
-			W: map[uint32]float64{1: 600, 2: 700, 3: 550}}, true
+			W: map[uint16]float64{1: 600, 2: 700, 3: 550}}, true
 	}
 	box := content.Box{URX: 200, URY: 200}
 	got, err := Page(context.Background(), []byte(cidTextOracleContent), box, 1, nil, fonts)
